First commit of Windows port of TinyEngine and examples.
TinyEngine itself is almost entirely unchanged from the original. The only major difference is a duplicate of TinyEngine.h placed in the win directory that adds some include statements that weren't required in the original.

Each example project has been turned into a Visual Studio project, all are contained in one Visual Studio solution. A template project has also been added that allows one to easily get a new running under Visual Studio.

All examples compile on Windows. Most are unchanged from the original, with the major exemption being 6_Tree that required a workaround for zero size vector causing problems with OpenGL on Debug builds.

Next task to complete is creating a README.md as there are a number of tasks that have to be completed by the user to set up their machine to be able to compile the examples.